+++ src/index.ts
@@ -7,3 +7,10 @@ export * from "./utils/errors";
 export { PostsService } from "./services/posts";
 
 export { HttpClient } from "./utils/http";
+
+export { extractStructuredData } from "./utils/structured-data";
+export type {
+  JsonLdObject,
+  StructuredDataOptions,
+  PostStructuredData,
+} from "./utils/structured-data";